--- a/src/container/index.ts
+++ b/src/container/index.ts
@@ -7,14 +7,90 @@ import { PostService } from '../services/post-service';
 
 /**
  * Dependency Injection Container
- * Provides all services as singleton instances
+ * Provides all services as singleton instances with async initialization
  */
-export const containerPlugin = new Elysia({ name: 'container' }).decorate('services', {
-  auth: new AuthService(prisma),
-  post: new PostService(prisma),
-  category: new CategoryService(prisma),
-  file: new FileService(prisma),
-});
+class ServiceContainerImpl {
+  private _auth: AuthService | null = null;
+  private _post: PostService | null = null;
+  private _category: CategoryService | null = null;
+  private _file: FileService | null = null;
+  private _initialized: boolean = false;
+
+  get isInitialized(): boolean {
+    return this._initialized;
+  }
+
+  async initialize(): Promise<void> {
+    if (this._initialized) {
+      return;
+    }
+
+    try {
+      // Create service instances
+      this._auth = new AuthService(prisma);
+      this._post = new PostService(prisma);
+      this._category = new CategoryService(prisma);
+      this._file = new FileService(prisma);
+
+      // Initialize services that need async setup
+      await Promise.all([
+        this._auth.initialize(),
+        this._post.initialize(),
+        this._category.initialize(),
+        this._file.initialize(),
+      ]);
+
+      this._initialized = true;
+      console.log('All services initialized successfully');
+    } catch (error) {
+      console.error('Failed to initialize services:', error);
+      throw error;
+    }
+  }
+
+  get auth(): AuthService {
+    if (!this._auth) {
+      throw new Error('Services not initialized. Call initialize() first.');
+    }
+    return this._auth;
+  }
+
+  get post(): PostService {
+    if (!this._post) {
+      throw new Error('Services not initialized. Call initialize() first.');
+    }
+    return this._post;
+  }
+
+  get category(): CategoryService {
+    if (!this._category) {
+      throw new Error('Services not initialized. Call initialize() first.');
+    }
+    return this._category;
+  }
+
+  get file(): FileService {
+    if (!this._file) {
+      throw new Error('Services not initialized. Call initialize() first.');
+    }
+    return this._file;
+  }
+}
+
+const serviceContainer = new ServiceContainerImpl();
+
+export const containerPlugin = new Elysia({ name: 'container' })
+  .decorate('services', serviceContainer)
+  .onStart(async () => {
+    // Initialize services when the server starts
+    await serviceContainer.initialize();
+  })
+  .onRequest(async () => {
+    // Ensure services are initialized before handling requests
+    if (!serviceContainer.isInitialized) {
+      await serviceContainer.initialize();
+    }
+  });
 
 // Type definition for service container
 export interface ServiceContainer {
@@ -22,11 +98,4 @@ export interface ServiceContainer {
   post: PostService;
   category: CategoryService;
   file: FileService;
-}
-
-// Extend Elysia context with services
-declare module 'elysia' {
-  interface Context {
-    services: ServiceContainer;
-  }
 }