--- a/src/lib/prisma.ts
+++ b/src/lib/prisma.ts
@@ -8,11 +8,18 @@ const prisma = new PrismaClient({
       url: process.env.DATABASE_URL,
     },
   },
+  // ログ設定
+  log:
+    process.env.NODE_ENV === 'development'
+      ? ['query', 'info', 'warn', 'error']
+      : process.env.NODE_ENV === 'test'
+        ? ['error']
+        : ['error'],
   // 接続プールの設定
-  log: process.env.NODE_ENV === 'development' ? ['query', 'info', 'warn', 'error'] : ['error'],
-  // 接続タイムアウトの設定
   transactionOptions: {
     timeout: 10000, // 10秒
+    maxWait: 5000, // 5秒
+    isolationLevel: 'Serializable',
   },
 });