… SECURITY_FIXES_REPORT.md
… __test__/auth.test.ts
… dev.db
… src/routes/posts.ts
… src/services/auth-service.ts
… src/services/post-service.ts
… src/types/auth.ts
… src/types/category.ts